Move the receive and intake tests into their own file
The receiving half: chunk assembly, record id normalization and its edge
cases, the dedup bookkeeping that rides with it, field reconstruction,
intake validation, and the two receiver-side diagnostics.
